requirements.txt
absl-py==1.4.0
altair==5.2.0
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.2.0
autopep8==2.0.1
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.28.3
colorama==0.4.6
comm==0.2.2
comtypes==1.1.14
contourpy==1.2.0
crispy-bootstrap4==2022.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-allauth==0.63.6
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-environ==0.9.0
django-htmx==1.18.0
django-jazzmin==3.0.0
django-redis==5.4.0
django-stubs==1.14.0
django-stubs-ext==0.7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib @ file:///C:/Users/aashi/Desktop/filesfiles/grapH/Dlib-library-Installation-main/Dlib-library-Installation-main/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
drf-yasg==1.21.7
environ==1.0
executing==1.2.0
Faker==25.4.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.16.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==7.0.1
inflection==0.5.1
ipykernel==6.29.5
ipython==8.11.0
jax==0.4.6
jedi==0.18.2
Jinja2==3.1.3
joblib==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.15.0
kiwisolver==1.4.5
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.5
mypy==1.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2022.12.19
platformdirs==2.6.2
playsound==1.3.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==7.0.0
psycopg2==2.9.5
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycodestyle==2.10.0
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.9.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.2.0
pytz==2022.7.1
pywin32==305
PyYAML==6.0
pyzmq==26.4.0
redis==4.6.0
referencing==0.33.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.9.0
sqlparse==0.4.3
stack-data==0.6.2
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
Theano==1.0.5
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.64.1
traitlets==5.9.0
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.6
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.13
validators==0.22.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
watchdog==4.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.17.0
